+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/network/CSPacketWorktableLockedModeToggle.java
@@ -1,5 +1,6 @@
 package com.codetaylor.mc.artisanworktables.modules.worktables.network;
 
+import com.codetaylor.mc.artisanworktables.modules.worktables.ModuleWorktablesConfig;
 import com.codetaylor.mc.artisanworktables.modules.worktables.tile.spi.TileEntityBase;
 import com.codetaylor.mc.athenaeum.spi.packet.SPacketTileEntityBase;
 import com.codetaylor.mc.athenaeum.util.BlockHelper;
@@ -8,34 +9,33 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 
-/**
- * Sent from the client to the server to signal a fluid destroy.
- */
-public class CSPacketWorktableCreateModeToggle
-    extends SPacketTileEntityBase<CSPacketWorktableCreateModeToggle> {
+public class CSPacketWorktableLockedModeToggle
+    extends SPacketTileEntityBase<CSPacketWorktableLockedModeToggle> {
 
   @SuppressWarnings("unused")
-  public CSPacketWorktableCreateModeToggle() {
+  public CSPacketWorktableLockedModeToggle() {
     // Serialization
   }
 
-  public CSPacketWorktableCreateModeToggle(BlockPos blockPos) {
+  public CSPacketWorktableLockedModeToggle(BlockPos blockPos) {
 
     super(blockPos);
   }
 
   @Override
   public IMessage onMessage(
-      CSPacketWorktableCreateModeToggle message,
+      CSPacketWorktableLockedModeToggle message,
       MessageContext ctx,
       TileEntity tileEntity
   ) {
 
     if (tileEntity instanceof TileEntityBase) {
       TileEntityBase table = (TileEntityBase) tileEntity;
-      table.setCreative(!table.isCreative());
-      CSPacketWorktableClear.clear(table, CSPacketWorktableClear.CLEAR_ALL);
-      BlockHelper.notifyBlockUpdate(table.getWorld(), table.getPos());
+
+      if (ModuleWorktablesConfig.allowSlotLockingForTier(table.getTier())) {
+        table.setLocked(!table.isLocked());
+        BlockHelper.notifyBlockUpdate(table.getWorld(), table.getPos());
+      }
     }
 
     return null;